Rename and move the account settings component
