+++ app/components/users/account_settings_component.html.erb
@@ -26,7 +26,7 @@
 
   <div class="flex flex-col gap-y-1 sm:flex-row sm:items-center sm:gap-x-3">
     <%= f.label :event_awareness, "How did you hear about us? 💡" %>
-    <%= f.select :event_awareness, UserFormComponent::AWARENESS_OPTIONS.invert.to_a, { prompt: "", selected: @user.event_awareness }, class: "input self-end pr-8 text-ellipsis" %>
+    <%= f.select :event_awareness, Users::AccountSettingsComponent::AWARENESS_OPTIONS.invert.to_a, { prompt: "", selected: @user.event_awareness }, class: "input self-end pr-8 text-ellipsis" %>
   </div>
 
   <%= f.submit "Save", class: "button" %>
