+++ app/views/pages/setup.html.erb
@@ -119,7 +119,7 @@
   </ol>
 </div>
 
-<%= render UserFormComponent.new(user: current_user) %>
+<%= render Users::AccountSettingsComponent.new(user: current_user) %>
 
 <p class="mt-4 italic text-sm opacity-80">
   If you have completed all these steps, <b>waited 10 minutes</b>, and your status remains
